requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
attrs==21.4.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.3
cryptography==37.0.2
Flask==2.1.2
flask-ngrok==0.0.25
frozenlist==1.3.0
idna==3.3
importlib-metadata==4.11.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
meraki==1.20.0
multidict==6.0.2
pycparser==2.21
python-dotenv==0.20.0
requests==2.27.1
urllib3==1.26.9
Werkzeug==2.1.2
yarl==1.7.2
zipp==3.8.0
